Kernel: Refactor interrupts and initialise APIC
Enable the APIC and set up a demo APIC timer interrupt. This update
configures the Local APIC on the boot processor, sets up the spurious
interrupt vector, and configures a timer with a dummy timer interrupt
handler. This will need re-working, but demonstrates that the basic
components can be made operational.