Kernel: Add section and helper macro to mark code blocks as user-mode
This is a quick-and-dirty hack to test the kernel's ability to jump into
user mode code. User mode code can be put in the USER_TEXT section,
which can be mapped into user-space memory easily. This hack will be
removed in the future when we can load real user-mode binaries from
disk.