Bootloader: Add graphical bootloader console
We're trying to get WorkbenchOS booting on real hardware, and thus can't
really use the bochs style debug port for getting output. I've written
up a basic line console which prints to the display using the graphics
mode and a hand-written bitmap font. It should work for now, even though
it's extremely limited.