Bootloader: Fix up paging structures so they don't fault on real HW
The kernel was page faulting while setting up the memory allocator. It
seemed to be caused by a reserved bit being set in one of the paging
structures. I solved this by zeroising the paging structures and
disabling the NX bit on the identity mapped PDPT (perhaps not supported
for this configuration on the laptop I'm testing?)